+++ projects/storefrontapp-e2e-cypress/cypress/helpers/customer-ticketing/customer-ticketing-helpers/create-tickets.ts
@@ -1,0 +1,70 @@
+
+import * as commons from './customer-ticketing-commons';
+import {
+  TestTicketDetails,
+} from './customer-ticketing-commons';
+
+export function openCreateTicketPopup() {
+  cy.get('cx-customer-ticketing-list').should('exist');
+  cy.get('cx-customer-ticketing-create button').click();
+  cy.get('cx-customer-ticketing-create-dialog').contains('Add New Request');
+}
+
+
+export function fillTicketDetails(ticketDetails: TestTicketDetails){
+  const CATEGORY_SELECT = 0;
+
+  cy.get('cx-customer-ticketing-create-dialog').within(() => {
+    cy.get('.cx-customer-ticket-form-container').should('contain', 'Add New Request');
+    cy.get('textarea').first().type(ticketDetails.subject);
+    cy.contains('.cx-customer-ticket-label', 'Category').get('select').eq(CATEGORY_SELECT).select(ticketDetails.category);
+    cy.get('textarea').last().type(ticketDetails.message);
+    cy.get('cx-form-errors').should('not.be.visible');
+  });
+}
+
+
+export function addFile(filename: string){
+  cy.get('cx-file-upload input[type="file"]').attachFile({filePath: '../helpers/customer-ticketing/files-to-upload/' + filename}, { allowEmpty: true });
+  cy.get('p').contains(filename);
+}
+
+export function clickSubmit(){
+  cy.contains('[type="button"]', 'Submit').click();
+}
+
+export function clickCancel(){
+  cy.contains('[type="button"]', 'Cancel').click();
+}
+
+export function clickClose(){
+  cy.get('cx-customer-ticketing-create-dialog').within(() => {
+    cy.get('.cx-customer-ticket-form-container').should('contain', 'Add New Request');
+    cy.get('button[aria-label="Close"]').click();
+  });
+}
+
+export function verifyFileTooLargeErrorIsShown(){
+  cy.get('cx-form-errors').should('be.visible');
+  cy.get('cx-form-errors').get('p').contains('File size should not exceed 10 MB');
+
+}
+
+
+export function verifyFieldValidationErrorShown(){
+  cy.get('cx-form-errors').should('be.visible');
+}
+
+
+export function createTicket(ticketDetails: TestTicketDetails){
+  openCreateTicketPopup();
+  fillTicketDetails(ticketDetails);
+  clickSubmit();
+  commons.verifyGlobalMessage();
+}
+
+export function createMultipleTickets(numberOfTicket: number, ticketDetails: TestTicketDetails){
+  for (let i = 0; i < numberOfTicket; i++){
+    createTicket(ticketDetails);
+  }
+}
